+++ src/components/CanvasPage/MenuSection/AIBackground.tsx
@@ -1,0 +1,3 @@
+export default function AIBackground() {
+	return <div>AIBackground</div>
+}
