+++ src/components/CanvasPage/MenuSection/AISticker.tsx
@@ -1,0 +1,3 @@
+export default function AISticker() {
+	return <div>AISticker</div>
+}
