+++ src/components/CanvasPage/MenuSection/Invite.tsx
@@ -1,0 +1,3 @@
+export default function Invite() {
+	return <div>Invite</div>
+}
