+++ src/components/CanvasPage/MenuSection/MyCanvas.tsx
@@ -1,0 +1,3 @@
+export default function MyCanvas() {
+	return <div>MyCanvas</div>
+}
