+++ src/components/CanvasPage/MenuSection/Upload.tsx
@@ -1,0 +1,3 @@
+export default function Upload() {
+	return <div>Upload</div>
+}
