+++ src/components/MainPage/CanvasPreview/index.tsx
